refactor(models,l10n): finish model purity — move localized labels to UI extensions
The remaining 6 models importing l10n (media_type, collection_sort_mode,
collection_list_sort_mode, search_sort, card_link, calendar_recurrence)
lose their localizedLabel-style methods to extensions under
shared/constants/*_ui.dart; cardSubcategoryLabel joins CollectionItemUi.
Dead SearchSortField.localizedShortLabel / localizedDisplayLabel deleted.
lib/shared/models/ is now fully pure Dart (no flutter / dart:ui / l10n).
