+++ lib/features/collections/screens/home_screen.dart
@@ -33,6 +33,7 @@ import '../widgets/edit_collection_dialog.dart';
 import '../widgets/import_progress_dialog.dart';
 import '../widgets/tag_management_dialog.dart';
 import 'collection_screen.dart';
+import '../../../shared/constants/collection_list_sort_mode_ui.dart';
 
 class HomeScreen extends ConsumerStatefulWidget {
   const HomeScreen({super.key});
