+++ lib/shared/constants/calendar_recurrence_ui.dart
@@ -1,0 +1,11 @@
+import '../../l10n/app_localizations.dart';
+import '../models/calendar_recurrence.dart';
+
+/// Presentation extras for [CalendarRecurrence].
+extension CalendarRecurrenceUi on CalendarRecurrence {
+  String localizedLabel(S l) => switch (this) {
+        CalendarRecurrence.once => l.recurrenceOnce,
+        CalendarRecurrence.weekly => l.recurrenceWeekly,
+        CalendarRecurrence.monthly => l.recurrenceMonthly,
+      };
+}
