+++ lib/shared/constants/collection_item_ui.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 
+import '../../l10n/app_localizations.dart';
 import '../models/collection_item.dart';
+import '../models/media_type.dart';
 import 'media_type_theme.dart';
 
 /// Presentation extras for [CollectionItem].
@@ -8,4 +10,19 @@ extension CollectionItemUi on CollectionItem {
   /// Custom items masquerading as another type get that type's icon.
   IconData get placeholderIcon =>
       MediaTypeTheme.placeholderIconFor(displayMediaType);
+
+  /// Qualifier shown after the media type: platform for games, movie/TV for
+  /// animation; `null` otherwise.
+  String? cardSubcategoryLabel(S l) {
+    switch (mediaType) {
+      case MediaType.game:
+        return platform == null ? null : platformName;
+      case MediaType.animation:
+        return platformId == AnimationSource.tvShow
+            ? l.mediaTypeTvShow
+            : l.mediaTypeMovie;
+      default:
+        return null;
+    }
+  }
 }
