+++ lib/shared/widgets/media_poster_card.dart
@@ -13,6 +13,7 @@ import '../theme/app_typography.dart';
 import 'cached_image.dart';
 import 'dual_rating_badge.dart';
 import '../constants/item_status_ui.dart';
+import '../constants/media_type_ui.dart';
 
 enum CardVariant {
   /// Full-size grid (collection + search).
